+++ .auxiliary/configuration/vulturefood.py
@@ -1,0 +1,22 @@
+ComparisonResult        # unused variable
+NominativeArguments     # unused variable
+PositionalArguments     # unused variable
+package_name            # unused variable
+
+# --- BEGIN: Injected by Copier ---
+# --- END: Injected by Copier ---
+
+install
+is_absence
+_typecheck_me
+AbsenceCell.from_optional
+AbsenceCell.extract
+AbsenceCell.extract_or
+AbsenceCell.extract_or_compute
+AbsenceCell.evaluate_or
+AbsenceCell.evaluate_or_true
+AbsenceCell.evaluate_or_false
+AbsenceCell.transform
+AbsenceCell.or_else
+AbsenceCell.to_optional
+adapt_dataclass